Fix recon routine

+++ src/reconlib.py
@@ -17,6 +17,7 @@ class ReconLib(BaseLib):
         self.tmp1 = f"tmp{self.id}1"
         self.tmp2 = f"tmp{self.id}2"
         self.tmp3 = f"tmp{self.id}3"
+        self.matching_key = ""
 
     def aggregate(self, cn, recon):
         """ aggregate and group imported data into temporary table """
@@ -25,9 +26,7 @@ class ReconLib(BaseLib):
         grouping_key = ""
         funcs = []
         sqllib = SqlLib()
-        if "Function" in recon:
-            funcs = recon["Function"]
-        field_list = sqllib.get_field_list(self.fields, self.types, funcs)
+        field_list = sqllib.get_fields(self.fields, self.types)
         sql = ""
         sql += f"insert into {self.tmp1} ({field_list}) "
         sql += f"select {field_list} from {self.tb1}"
@@ -44,52 +43,56 @@ class ReconLib(BaseLib):
         self.method = "reconlib.match()"
         sqllib = SqlLib()
         rule = recon["Rule"]
-        fields_key = recon["Key"]
-        matching_key = sqllib.get_sql_key(self.tmp1, self.tmp2, fields_key)
+        matching_key = sqllib.get_sql_key(self.tmp1, self.tmp2, recon["Fields"])
         cn.execute(f"update {self.tmp1} set id_parent = (select id from {self.tmp2} where 1=1 {matching_key})")
         cn.execute(f"update {self.tmp2} set id_parent = (select id from {self.tmp1} where 1=1 {matching_key})")
         cn.execute(f"update {self.tmp1} set id_parent = 0 where id_parent is null")
         cn.execute(f"update {self.tmp2} set id_parent = 0 where id_parent is null")
         cn.execute(f"update {self.tmp1} set recon='{self.name}', rule='{rule}', status='matched' where id_parent <> 0")
         cn.execute(f"update {self.tmp2} set recon='{self.name}', rule='{rule}', status='matched' where id_parent <> 0")
-        print(1)
+        self.matching_key = matching_key
+
     def compare(self, cn, recon):
         """ compare the records and relegate the status """
         self.method = "reconlib.match()"
         sql = ""
+        fields_key = ""
         utillib = UtilLib()
         sqllib = SqlLib()
         rule = recon["Rule"]
-        fields_key = recon["Key"]
-        fields_compare = recon["Compare"]
-        matching_key = sqllib.get_sql_key(self.tmp1, self.tmp2, fields_key)
-        fields_key = [(f"{self.tmp1}.{field}") for field in fields_key]
-        fields_key = sqllib.get_field_list(fields_key)
+        matching_key = self.matching_key
+        """ create temp table  to compare fields """
+        for field in recon["Fields"]:
+            if str(field["Type"]).strip().lower() == "key":
+                field_name = str(field["Name"]).strip().lower()
+                fields_key += f"{self.tmp1}.{field_name}, "
+        fields_key = fields_key.strip()[:-1].lower()
         """ keep difference and status in tmp3 """
         tablename = self.tmp3
-        for field in fields_compare:
-            cn.execute(f"drop table if exists {tablename}")
-            field = field.lower()
-            sql = ""
-            tmp1 = f"{self.tmp1}.{field}"
-            tmp2 = f"{self.tmp2}.{field}"
-            sql += f" create table {tablename} as"
-            sql += f" select"
-            sql += f" {fields_key}"
-            sql += f", ({tmp1} || '/' || {tmp2}) difference"
-            sql += f", ({tmp1} = {tmp2}) equality"
-            sql += f" from {self.tmp1}, {self.tmp2}"
-            sql += f" where {self.tmp1}.status = 'matched'"
-            sql += f" {matching_key}"
-            cn.execute(sql)
-            """ stamp the differences in tmp1/tmp2 tables """
-            for side in range(1,3):
-                temps = self.tmp1 if side == 1 else self.tmp2
-                matching_key = sqllib.get_sql_key(temps, self.tmp3, recon["Key"])
-                cn.execute(f"alter table {temps} add {field}_diff text default ''")
-                cn.execute(f"update {temps} set {field}_diff = (select difference from {self.tmp3} where equality = 0 {matching_key})")
-                cn.execute(f"update {temps} set {field}_diff = '' where {field}_diff is null")
-                cn.execute(f"update {temps} set status = 'divergent' where {field}_diff <> ''")
+        for field in recon["Fields"]:
+            if str(field["Type"]).strip().lower() == "compare":
+                cn.execute(f"drop table if exists {tablename}")
+                field = field["Name"].lower()
+                sql = ""
+                tmp1 = f"{self.tmp1}.{field}"
+                tmp2 = f"{self.tmp2}.{field}"
+                sql += f" create table {tablename} as"
+                sql += f" select"
+                sql += f" {fields_key}"
+                sql += f", ({tmp1} || '/' || {tmp2}) difference"
+                sql += f", ({tmp1} = {tmp2}) equality"
+                sql += f" from {self.tmp1}, {self.tmp2}"
+                sql += f" where {self.tmp1}.status = 'matched'"
+                sql += f" {matching_key}"
+                cn.execute(sql)
+                """ stamp the differences in tmp1/tmp2 tables """
+                for side in range(1,3):
+                    temps = self.tmp1 if side == 1 else self.tmp2
+                    matching_key = sqllib.get_sql_key(temps, self.tmp3, recon["Fields"])
+                    cn.execute(f"alter table {temps} add {field}_diff text default ''")
+                    cn.execute(f"update {temps} set {field}_diff = (select difference from {self.tmp3} where equality = 0 {matching_key})")
+                    cn.execute(f"update {temps} set {field}_diff = '' where {field}_diff is null")
+                    cn.execute(f"update {temps} set status = 'divergent' where {field}_diff <> ''")
 
     def stamp(self, cn, recon):
         """ update the final status from grouped tmp table to flat table """
@@ -97,10 +100,9 @@ class ReconLib(BaseLib):
         utillib = UtilLib()
         sqllib = SqlLib()
         rule = recon["Rule"]
-        fields_key = recon["Key"]
         match_info = ["id_parent", "recon", "rule", "status"]
-        matching_key1 = sqllib.get_sql_key(self.tb1, self.tmp1, fields_key)
-        matching_key2 = sqllib.get_sql_key(self.tb2, self.tmp2, fields_key)
+        matching_key1 = sqllib.get_sql_key(self.tb1, self.tmp1, recon["Fields"])
+        matching_key2 = sqllib.get_sql_key(self.tb2, self.tmp2, recon["Fields"])
         for field in match_info:
             cn.execute(f"update {self.tb1} set {field} = (select {field} from {self.tmp1} where 1=1 {matching_key1})")
             cn.execute(f"update {self.tb2} set {field} = (select {field} from {self.tmp2} where 1=1 {matching_key2})")
@@ -121,6 +123,8 @@ class ReconLib(BaseLib):
             utillib.log(message)
             self.logger.error(message)
         except BaseException as err:
-            self.logger.error(f"{self.method}: General error: {str(err)}")
+            message = f"{self.method}: General error -> {str(err)}"
+            utillib.log(message)
+            self.logger.error(message)
         finally:
             self.logger.info(f"{self.method}: Done")
